slack_bot/requirements.txt
pandas ==0.23.4
sqlalchemy == 1.2.12
slackclient
nest_asyncio
psycopg2-binary
